Changed panstamp ok from 0 to 1
Testing with water level sensor disconnected, about 10% of the time the
sensor status byte said the level sensor was online.  There are a lot
of checksum errors on the I2C - all the data was zero except checksum.
Zero is a valid value for the panstamp being online, so I changed it to
a 1.